+++ test/unit/configuration_test.rb
@@ -103,6 +103,7 @@ class ConfigurationTest < Test::Unit::TestCase
     assert_config_overridable :notifier_version
     assert_config_overridable :notifier_name
     assert_config_overridable :notifier_url
+    assert_config_overridable :delivery_method
     assert_config_overridable :environment_name
     assert_config_overridable :logger
     assert_config_overridable :source_extract_radius
@@ -119,13 +120,15 @@ class ConfigurationTest < Test::Unit::TestCase
   should "act like a hash" do
     config = Honeybadger::Configuration.new
     hash = config.to_hash
-    [:api_key, :backtrace_filters, :development_environments,
-     :environment_name, :host, :http_open_timeout, :http_read_timeout, :ignore,
-     :ignore_by_filters, :ignore_user_agent, :notifier_name, :notifier_url,
-     :notifier_version, :params_filters, :project_root, :port, :protocol,
-     :proxy_host, :proxy_pass, :proxy_port, :proxy_user, :secure,
-     :source_extract_radius, :async, :send_request_session, :debug, :fingerprint].each do |option|
-       assert_equal config[option], hash[option], "Wrong value for #{option}"
+    [ :api_key, :backtrace_filters, :development_environments,
+      :delivery_method, :environment_name, :host, :http_open_timeout,
+      :http_read_timeout, :ignore, :ignore_by_filters, :ignore_user_agent,
+      :notifier_name, :notifier_url, :notifier_version, :params_filters,
+      :project_root, :port, :protocol, :proxy_host, :proxy_pass, :proxy_port,
+      :proxy_user, :secure, :source_extract_radius, :async,
+      :send_request_session, :debug
+    ].each do |option|
+      assert_equal config[option], hash[option], "Wrong value for #{option}"
     end
   end
 
@@ -175,9 +178,14 @@ class ConfigurationTest < Test::Unit::TestCase
     assert_replaces(:ignore_user_agent, :ignore_user_agent_only=)
   end
 
-  should "use development and test as development environments by default" do
+  should "use development as development environments by default" do
     config = Honeybadger::Configuration.new
     assert_same_elements %w(development test cucumber), config.development_environments
+  end
+
+  should "have test and cucumber as test environments" do
+    config = Honeybadger::Configuration.new
+    assert_same_elements %w(test cucumber), config.test_environments
   end
 
   should "be public in a public environment" do
@@ -191,6 +199,20 @@ class ConfigurationTest < Test::Unit::TestCase
     config = Honeybadger::Configuration.new
     config.development_environments = %w(staging)
     config.environment_name = 'staging'
+    assert !config.public?
+  end
+
+  should "not be public in a test environment" do
+    config = Honeybadger::Configuration.new
+    config.development_environments = []
+    config.environment_name = 'test'
+    assert !config.public?
+  end
+
+  should "not be public in a cucumber environment" do
+    config = Honeybadger::Configuration.new
+    config.development_environments = []
+    config.environment_name = 'cucumber'
     assert !config.public?
   end